move lcd serve command to a dedicated file
